Close peer connection and Multi-viewer support (#42)
* PeerConnection module changes

* mbedtls and main.c changes

* ice_controller and app_media_source changes

* Setting iceServersCount to 0

* Remove socket mutex deletion in destroy.

* Removing extra logs

* Remove unnecessary include

* Move skb buffer check to media port layer. Start/Stop media with mutex lock protect.

* Review suggestions.

* Change include projdef.h to FreeRTOS.h

---------

+++ examples/peer_connection/peer_connection_jitter_buffer.c
@@ -33,7 +33,7 @@ static void DiscardPackets( PeerConnectionJitterBuffer_t * pJitterBuffer,
     if( pJitterBuffer != NULL )
     {
         /* Free from start sequence to end sequence number. */
-        for( i = startSeq; i != endSeq + 1; i++ )
+        for( i = startSeq; i != ( uint16_t )( endSeq + 1 ); i++ )
         {
             index = PEER_CONNECTION_JITTER_BUFFER_WRAP( i,
                                                         PEER_CONNECTION_JITTER_BUFFER_MAX_ENTRY_NUM );
@@ -81,7 +81,7 @@ static PeerConnectionResult_t ParseFramesInJitterBuffer( PeerConnectionJitterBuf
         earliestBufferTimestamp = pJitterBuffer->newestReceivedTimestamp - pJitterBuffer->tolerenceRtpTimeStamp;
         /* Note that newest sequence is probably less than oldest sequence due to wrapping. */
         prev = pJitterBuffer->oldestReceivedSequenceNumber;
-        for( i = pJitterBuffer->oldestReceivedSequenceNumber; i != pJitterBuffer->newestReceivedSequenceNumber + 1; i++ )
+        for( i = pJitterBuffer->oldestReceivedSequenceNumber; i != ( uint16_t )( pJitterBuffer->newestReceivedSequenceNumber + 1 ); i++ )
         {
             index = PEER_CONNECTION_JITTER_BUFFER_WRAP( i,
                                                         PEER_CONNECTION_JITTER_BUFFER_MAX_ENTRY_NUM );
